microupd 5.3.0 Создать метод delete для односвязного списка[#84101]

+++ chapter_003/src/test/java/ru/job4j/arrays/SimpleArrayListTest.java
@@ -34,8 +34,8 @@ public class SimpleArrayListTest {
 
     @Test
     public void whenAddThreeElementsThenDeleteTwoOfThem() {
-        assertThat((Integer) list.delete(), is(3));
-        assertThat((Integer) list.delete(), is(2));
+        assertThat(list.delete(), is(3));
+        assertThat(list.delete(), is(2));
         assertThat(list.getSize(), is(1));
     }
 }
